Move files in with other rando files
