+++ soh/src/overlays/actors/ovl_En_Ds/z_en_ds.c
@@ -6,7 +6,7 @@
 
 #include "z_en_ds.h"
 #include "objects/object_ds/object_ds.h"
-#include "z64adult_trade_shuffle.h"
+#include "soh/Enhancements/randomizer/adult_trade_shuffle.h"
 
 #define FLAGS (ACTOR_FLAG_0 | ACTOR_FLAG_3)
