+++ soh/src/overlays/actors/ovl_En_Kz/z_en_kz.c
@@ -6,7 +6,7 @@
 
 #include "z_en_kz.h"
 #include "objects/object_kz/object_kz.h"
-#include "z64adult_trade_shuffle.h"
+#include "soh/Enhancements/randomizer/adult_trade_shuffle.h"
 
 #define FLAGS (ACTOR_FLAG_0 | ACTOR_FLAG_3)
